deprecatedになっていたメニュー周りの修正
