+++ AndroidApp/app/src/main/java/me/nya_n/notificationnotifier/views/screen/top/TopViewModel.kt
@@ -1,5 +1,7 @@
 package me.nya_n.notificationnotifier.views.screen.top
 
+import android.content.Context
+import android.net.Uri
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
@@ -8,15 +10,15 @@ import kotlinx.coroutines.launch
 import me.nya_n.notificationnotifier.R
 import me.nya_n.notificationnotifier.domain.entities.AppException
 import me.nya_n.notificationnotifier.domain.entities.Message
-import me.nya_n.notificationnotifier.domain.usecase.LoadAddressUseCase
-import me.nya_n.notificationnotifier.domain.usecase.NotifyTestUseCase
-import me.nya_n.notificationnotifier.domain.usecase.SaveAddressUseCase
+import me.nya_n.notificationnotifier.domain.usecase.*
 import me.nya_n.notificationnotifier.utils.Event
 
 class TopViewModel(
     loadAddressUseCase: LoadAddressUseCase,
     private val saveAddressUseCase: SaveAddressUseCase,
     private val notifyTestUseCase: NotifyTestUseCase,
+    private val importDataUseCase: ImportDataUseCase,
+    private val exportDataUseCase: ExportDataUseCase,
 ) : ViewModel() {
     private val _message = MutableLiveData<Event<Message>>()
     val message: LiveData<Event<Message>> = _message
@@ -53,6 +55,38 @@ class TopViewModel(
                 .onFailure {
                     it.printStackTrace()
                     _message.postValue(Event(Message.Error(R.string.notify_test_failed)))
+                }
+        }
+    }
+
+    /**
+     * バックアップのために外部ストレージにデータを保存
+     */
+    fun exportData(context: Context, uri: Uri) {
+        viewModelScope.launch {
+            exportDataUseCase(context, uri)
+                .onSuccess {
+                    _message.postValue(Event(Message.Notice(R.string.export_succeeded)))
+                }
+                .onFailure {
+                    it.printStackTrace()
+                    _message.postValue(Event(Message.Error(R.string.export_failed)))
+                }
+        }
+    }
+
+    /**
+     * 外部ストレージのバックアップからデータを復元
+     */
+    fun importData(context: Context, uri: Uri) {
+        viewModelScope.launch {
+            importDataUseCase(context, uri)
+                .onSuccess {
+                    _message.postValue(Event(Message.Notice(R.string.import_succeeded)))
+                }
+                .onFailure {
+                    it.printStackTrace()
+                    _message.postValue(Event(Message.Error(R.string.import_failed)))
                 }
         }
     }
